Move chalk-solve to main Chalk crate
