+++ Cargo.toml
@@ -12,16 +12,32 @@ edition = "2018"
 [features]
 bench = []
 
-[dependencies.chalk-repl]
-version = "0.1.0"
-path = "chalk-repl"
+[dependencies]
+derive-new = "0.5.6"
+ena = "0.10.1"
+failure = "0.1"
+itertools = "0.7.8"
+petgraph = "0.4.13"
+rustc-hash = { version = "1.0.0" }
 
-[dependencies.chalk-solve]
+[dependencies.chalk-macros]
 version = "0.1.0"
-path = "chalk-solve"
+path = "chalk-macros"
 
-[dependencies.chalk-tests]
+[dependencies.chalk-engine]
+version = "0.9.0"
+path = "chalk-engine"
+
+[dependencies.chalk-ir]
 version = "0.1.0"
-path = "chalk-tests"
+path = "chalk-ir"
+
+[dependencies.chalk-rust-ir]
+version = "0.1.0"
+path = "chalk-rust-ir"
 
 [workspace]
+members = [
+    "chalk-repl",
+    "chalk-tests",
+]
