+++ chalk-integration/src/db.rs
@@ -2,6 +2,9 @@ use crate::error::ChalkError;
 use crate::lowering::LowerGoal;
 use crate::program::Program;
 use crate::query::{Lowering, LoweringDatabase};
+use chalk::RustIrDatabase;
+use chalk::Solution;
+use chalk::SolverChoice;
 use chalk_ir::tls;
 use chalk_ir::Goal;
 use chalk_ir::Identifier;
@@ -21,9 +24,6 @@ use chalk_rust_ir::AssociatedTyDatum;
 use chalk_rust_ir::ImplDatum;
 use chalk_rust_ir::StructDatum;
 use chalk_rust_ir::TraitDatum;
-use chalk_solve::RustIrDatabase;
-use chalk_solve::Solution;
-use chalk_solve::SolverChoice;
 use salsa::Database;
 use std::sync::Arc;
