+++ chalk-integration/src/lib.rs
@@ -1,0 +1,12 @@
+#[macro_use]
+extern crate chalk_macros;
+#[macro_use]
+extern crate failure;
+
+pub mod db;
+pub mod error;
+pub mod lowering;
+pub mod program;
+pub mod program_environment;
+pub mod query;
+pub mod wf;
