+++ chalk-tests/src/lib.rs
@@ -7,11 +7,4 @@ extern crate failure;
 #[macro_use]
 mod test_util;
 
-pub mod db;
-pub mod error;
-pub mod lowering;
-pub mod program;
-pub mod program_environment;
-pub mod query;
-
 mod test;
