+++ chalk-tests/src/test.rs
@@ -1,10 +1,10 @@
 #![cfg(test)]
 
-use crate::db::ChalkDatabase;
-use crate::query::LoweringDatabase;
+use chalk::ext::*;
+use chalk::{Solution, SolverChoice};
+use chalk_integration::db::ChalkDatabase;
+use chalk_integration::query::LoweringDatabase;
 use chalk_ir;
-use chalk_solve::ext::*;
-use chalk_solve::{Solution, SolverChoice};
 
 #[cfg(feature = "bench")]
 mod bench;
