+++ chalk-tests/src/test/bench.rs
@@ -3,9 +3,9 @@
 extern crate test;
 use self::test::Bencher;
 
+use chalk::ext::*;
+use chalk::SolverChoice;
 use chalk_ir;
-use chalk_solve::ext::*;
-use chalk_solve::SolverChoice;
 
 use super::{assert_result, parse_and_lower_goal, parse_and_lower_program_with_env};
