+++ chalk-tests/src/test/slg.rs
@@ -1,9 +1,9 @@
 #![cfg(test)]
 
-use crate::db::ChalkDatabase;
 use crate::test_util::*;
-use chalk_solve::ext::*;
-use chalk_solve::SolverChoice;
+use chalk::ext::*;
+use chalk::SolverChoice;
+use chalk_integration::db::ChalkDatabase;
 
 macro_rules! test {
     (program $program:tt $(goal $goal:tt first $n:tt with max $depth:tt { $expected:expr })*) => {
