+++ chalk-tests/src/test_util.rs
@@ -9,9 +9,9 @@ macro_rules! lowering_success {
         let program_text = stringify!($program);
         assert!(program_text.starts_with("{"));
         assert!(program_text.ends_with("}"));
-        let result = crate::db::ChalkDatabase::with(
+        let result = chalk_integration::db::ChalkDatabase::with(
             &program_text[1..program_text.len() - 1],
-            chalk_solve::SolverChoice::default(),
+            chalk::SolverChoice::default(),
         )
         .checked_program();
         if let Err(ref e) = result {
@@ -26,9 +26,9 @@ macro_rules! lowering_error {
         let program_text = stringify!($program);
         assert!(program_text.starts_with("{"));
         assert!(program_text.ends_with("}"));
-        let error = crate::db::ChalkDatabase::with(
+        let error = chalk_integration::db::ChalkDatabase::with(
             &program_text[1..program_text.len() - 1],
-            chalk_solve::SolverChoice::default(),
+            chalk::SolverChoice::default(),
         )
         .checked_program()
         .unwrap_err();
