+++ src/coinductive_goal.rs
