+++ src/infer.rs
