+++ src/infer/invert.rs
