+++ src/infer/var.rs
