+++ src/solve/slg/resolvent.rs
